change id columns in batch_job_sep from int to bigint

+++ ddl/dw/dimensions/dwh_dim_batch_job_sep.sql
@@ -20,7 +20,7 @@ DROP TABLE IF EXISTS `dwh_dim_batch_job_sep`;
 
 CREATE TABLE `dwh_dim_batch_job_sep` (
   `dwh_id` bigint NOT NULL AUTO_INCREMENT,
-  `id` INT(11) NOT NULL,
+  `id` bigint NOT NULL,
   `job_type_id` INT(6) DEFAULT NULL,
   `job_sub_type_id` INT(6) DEFAULT NULL,
   `object_id` VARCHAR(20) DEFAULT NULL,
